Services updates + order customer service  + first controller

+++ src/main/java/com/PersonalProject/Jemo/repository/MvtStkRepository.java
@@ -3,5 +3,5 @@ package com.PersonalProject.Jemo.repository;
 import com.PersonalProject.Jemo.model.MvtStk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MvtStkRepository extends JpaRepository<MvtStk, Integer> {
+public interface MvtStkRepository extends JpaRepository<MvtStk, Long> {
 }
