+++ src/main/java/com/PersonalProject/Jemo/repository/ProductRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     Optional<Product> findByName(String productName);
-    List<Product> findAllByCategoryId(Integer id);
+    List<Product> findAllByCategoryId(Long id);
 }
