+++ src/main/java/com/PersonalProject/Jemo/repository/SupplierRepository.java
@@ -3,5 +3,5 @@ package com.PersonalProject.Jemo.repository;
 import com.PersonalProject.Jemo.model.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
+public interface SupplierRepository extends JpaRepository<Supplier, Long> {
 }
