+++ src/main/java/com/PersonalProject/Jemo/services/SupplierService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface SupplierService {
 
     SupplierDto save(SupplierDto supplierDto);
-    SupplierDto findById(Integer id);
+    SupplierDto findById(Long id);
     List<SupplierDto> findAll();
-    void delete(Integer id);
+    void delete(Long id);
 }
