+++ src/main/java/com/PersonalProject/Jemo/services/UserService.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface UserService {
 
     UserDto save(UserDto userDto);
-    UserDto findById(Integer id);
+    UserDto findById(Long id);
     List<UserDto> findAll();
-    void delete(Integer id);
+    void delete(Long id);
     UserDto findByEmail(String email);
     UserDto ChangePassword(ModifyPasswordDto modifyPasswordDto);
 }
